Update Master with working balloon and command board test code

+++ UNITE_Main.X/adc1.h
@@ -50,5 +50,5 @@ typedef enum
 void InitializeADC1(void);
 
 uint16_t ADC1_ResultGet(uint16_t *buffer);
-int ADC1_ResultGetFromChannel(int);
+uint8_t ADC1_ResultGetFromChannel(int);
 void ADC1_ManualSampleForSec(unsigned long);
